CTCTRADERS-2344 Move test only models to test only package
