+++ app/api/controllers/testOnly/models/ControlResult.scala
@@ -14,22 +14,17 @@
  * limitations under the License.
  */
 
-package api.models
+package api.controllers.testOnly.models
 
-import play.api.libs.functional.syntax._
 import play.api.libs.json.Json
 import play.api.libs.json.OWrites
 import play.api.libs.json.Reads
-import play.api.libs.json.__
 
-case class OfficeOfTransit(id: String, name: String)
+case class ControlResult(code: String, description: String)
 
-object OfficeOfTransit {
+object ControlResult {
 
-  implicit val writes: OWrites[OfficeOfTransit] = Json.writes[OfficeOfTransit]
+  implicit val writes: OWrites[ControlResult] = Json.writes[ControlResult]
 
-  implicit val readFromFile: Reads[OfficeOfTransit] =
-    ((__ \ "ID").read[String] and
-      (__ \ "NAME").read[String])(OfficeOfTransit.apply _)
-
+  implicit val readFromFile: Reads[ControlResult] = Json.reads[ControlResult]
 }
