+++ app/api/controllers/testOnly/models/DocumentType.scala
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package api.models
+package api.controllers.testOnly.models
 
 import play.api.libs.json.Json
 import play.api.libs.json.OWrites
 import play.api.libs.json.Reads
 
-case class KindOfPackage(code: String, description: String)
+final case class DocumentType(code: String, description: String, transportDocument: Boolean)
 
-object KindOfPackage {
+object DocumentType {
 
-  implicit val writes: OWrites[KindOfPackage] = Json.writes[KindOfPackage]
+  implicit val writes: OWrites[DocumentType] = Json.writes[DocumentType]
 
-  implicit val readFromFile: Reads[KindOfPackage] = Json.reads[KindOfPackage]
+  implicit val readFromFile: Reads[DocumentType] = Json.reads[DocumentType]
 
 }
