+++ app/api/controllers/testOnly/models/PreviousDocumentType.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package api.models
+package api.controllers.testOnly.models
 
 import play.api.libs.json.Json
 import play.api.libs.json.OWrites
