+++ app/api/controllers/testOnly/services/DangerousGoodsCodeService.scala
@@ -16,7 +16,7 @@
 
 package api.controllers.testOnly.services
 
-import api.models.DangerousGoodsCode
+import api.controllers.testOnly.models.DangerousGoodsCode
 import javax.inject.Inject
 import play.api.Environment
