+++ app/api/controllers/testOnly/services/PreviousDocumentTypeService.scala
@@ -16,7 +16,7 @@
 
 package api.controllers.testOnly.services
 
-import api.models.PreviousDocumentType
+import api.controllers.testOnly.models.PreviousDocumentType
 import javax.inject.Inject
 import play.api.Environment
