+++ test/api/controllers/testOnly/models/OfficeOfTransitSpec.scala
@@ -14,38 +14,38 @@
  * limitations under the License.
  */
 
-package api.models
+package api.controllers.testOnly.models
 
 import api.generators.ModelGenerators
 import base.SpecBase
 import org.scalacheck.Arbitrary.arbitrary
-import org.scalacheck.Gen
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
+import org.scalacheck.Gen
 import play.api.libs.json._
 
-class DangerousGoodsCodeSpec extends SpecBase with ScalaCheckPropertyChecks with ModelGenerators {
+class OfficeOfTransitSpec extends SpecBase with ScalaCheckPropertyChecks with ModelGenerators {
 
-  private val dangerousGoodsCode = DangerousGoodsCode("0004", "AMMONIUM PICRATE dry or wetted with less than 10% water, by mass")
+  private val officesOfTransit = OfficeOfTransit("DE009583", "Stuttgart, Stuttgart-Hauptbahnhof")
 
   "Transit Office model" - {
 
     "must deserialise from json to a valid model" in {
 
-      val result = validDangerousGoodsCodeJson(dangerousGoodsCode).as[DangerousGoodsCode]
+      val result = validOfficeOfTransitsJson(officesOfTransit).as[OfficeOfTransit]
 
-      result mustBe dangerousGoodsCode
+      result mustBe officesOfTransit
     }
 
     "must serialize to json from valid json model" in {
-      val result = Json.toJson(dangerousGoodsCode)
-      result mustBe expectedDangerousGoodsCodeJson()
+      val result = Json.toJson(officesOfTransit)
+      result mustBe expectedOfficesOfTransitJson()
     }
 
     "must serialize to json and deserialize to a valid model" in {
-      forAll(arbitrary[DangerousGoodsCode]) {
-        dangerousGoodsCode =>
-          val json = validDangerousGoodsCodeJson(dangerousGoodsCode)
-          json.as[DangerousGoodsCode] mustBe dangerousGoodsCode
+      forAll(arbitrary[OfficeOfTransit]) {
+        officeOfTransit =>
+          val json = validOfficeOfTransitsJson(officeOfTransit)
+          json.as[OfficeOfTransit] mustBe officeOfTransit
       }
     }
 
@@ -58,22 +58,22 @@ class DangerousGoodsCodeSpec extends SpecBase with ScalaCheckPropertyChecks with
 
       forAll(invalidJsonGenerator) {
         invalidJson =>
-          intercept[JsResultException](invalidJson.as[DangerousGoodsCode])
+          intercept[JsResultException](invalidJson.as[OfficeOfTransit])
       }
     }
   }
 
-  def expectedDangerousGoodsCodeJson(dangerousGoodsCode: DangerousGoodsCode = dangerousGoodsCode): JsValue =
+  def expectedOfficesOfTransitJson(officeOfTransit: OfficeOfTransit = officesOfTransit): JsValue =
     Json.obj(
-      "code"        -> dangerousGoodsCode.code,
-      "description" -> dangerousGoodsCode.description
+      "id"   -> officeOfTransit.id,
+      "name" -> officeOfTransit.name
     )
 
-  def validDangerousGoodsCodeJson(dangerousGoodsCode: DangerousGoodsCode): JsValue =
+  def validOfficeOfTransitsJson(office: OfficeOfTransit): JsValue =
     Json.parse(s"""
                   |{
-                  |  "code":"${dangerousGoodsCode.code}",
-                  |  "description":"${dangerousGoodsCode.description}"
+                  |  "ID":"${office.id}",
+                  |  "NAME":"${office.name}"
                   |  }
                   |""".stripMargin)
