+++ test/api/controllers/testOnly/services/DocumentTypeServiceSpec.scala
@@ -16,7 +16,7 @@
 
 package api.controllers.testOnly.services
 
-import api.models.DocumentType
+import api.controllers.testOnly.models.DocumentType
 import base.SpecBaseWithAppPerSuite
 
 class DocumentTypeServiceSpec extends SpecBaseWithAppPerSuite {
